+++ lib/weather/weather_screen.dart
@@ -6,24 +6,10 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_training/weather/location.dart';
 import 'package:flutter_training/weather/weather.dart';
 import 'package:flutter_training/weather/weather_alert_dialog.dart';
+import 'package:flutter_training/weather/weather_notifier.dart';
 import 'package:flutter_training/weather/weather_panel.dart';
 import 'package:json_annotation/json_annotation.dart';
-import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:yumemi_weather/yumemi_weather.dart';
-
-part 'weather_screen.g.dart';
-
-@riverpod
-class WeatherNotifier extends _$WeatherNotifier {
-  @override
-  Weather? build() {
-    return null;
-  }
-
-  void update(Weather weather) {
-    state = weather;
-  }
-}
 
 class WeatherScreen extends ConsumerStatefulWidget {
   const WeatherScreen({super.key});
